fix(gpt-gateway): issue related to wrong deserialisation of kafka message on conversation start
