+++ libs/schema-registry/src/index.ts
@@ -10,6 +10,7 @@ export * as CanUserAccessBuild from "./lib/can-user-access-build";
 export * as DBSchemaImportRequest from "./lib/db-schema-import-request";
 export * as UserActionLog from "./lib/user-action-log";
 export * as UserAction from "./lib/user-action";
+export * as AiConversationStart from "./lib/ai-conversion-start";
 export * as AiConversationComplete from "./lib/ai-conversion-complete";
 export * as UserBuild from "./lib/user-build";
