+++ libs/schema-registry/src/lib/ai-conversion-start/key.ts
@@ -1,0 +1,6 @@
+import { IsString } from "class-validator";
+
+export class Key {
+  @IsString()
+  requestUniqueId!: string;
+}
