+++ libs/schema-registry/src/lib/ai-conversion-start/value.ts
@@ -1,0 +1,18 @@
+import { IsString } from "class-validator";
+
+class MessageParam {
+  @IsString()
+  name!: string;
+
+  @IsString()
+  value!: string;
+}
+
+export class Value {
+  @IsString()
+  requestUniqueId!: string;
+  @IsString()
+  messageTypeKey!: string;
+
+  params!: MessageParam[];
+}
