--- packages/gpt-gateway/src/dto/MessageParam.ts
@@ -1,16 +1,0 @@
-import { Field, InputType } from "@nestjs/graphql";
-
-@InputType()
-class MessageParam {
-  @Field(() => String, {
-    nullable: false,
-  })
-  name!: string;
-
-  @Field(() => String, {
-    nullable: false,
-  })
-  value!: string;
-}
-
-export { MessageParam };
