+++ packages/gpt-gateway/src/kafka/kafka.controller.ts
@@ -1,5 +1,4 @@
 import { Controller, Inject } from "@nestjs/common";
-import { StartConversationInput } from "../dto/StartConversationInput";
 import { ConversationTypeService } from "../conversationType/conversationType.service";
 import { AmplicationLogger } from "@amplication/util/nestjs/logging";
 import {
@@ -8,6 +7,8 @@ import {
   KafkaContext,
   Payload,
 } from "@nestjs/microservices";
+import { plainToInstance } from "class-transformer";
+import { AiConversationStart } from "@amplication/schema-registry";
 
 @Controller("kafka-controller")
 export class KafkaController {
@@ -20,11 +21,11 @@ export class KafkaController {
   @EventPattern("ai.conversation.start.1")
   async onAiConversationStart_1(
     @Payload()
-    value: string | Record<string, any> | null,
+    value: AiConversationStart.Value,
     @Ctx()
     context: KafkaContext
   ): Promise<void> {
-    const messageInput: StartConversationInput = JSON.parse(value.toString());
+    const messageInput = plainToInstance(AiConversationStart.Value, value);
 
     this.logger.info(`Got a new Gpt Conversation request item from queue.`, {
       requestedId: messageInput.requestUniqueId,
