--- packages/gpt-gateway/src/template/dto/ProcessTemplateArgs.ts
@@ -1,10 +1,0 @@
-import { ArgsType, Field } from "@nestjs/graphql";
-import { ProcessTemplateInput } from "./ProcessTemplateInput";
-
-@ArgsType()
-class ProcessTemplateArgs {
-  @Field(() => ProcessTemplateInput, { nullable: false })
-  data!: ProcessTemplateInput;
-}
-
-export { ProcessTemplateArgs };
